requirements.txt
streamlit
langchain-google-genai
google-generativeai
pypdf
chromadb
langchain
langchain-community
langchain-chroma